moved mc_nearest_facet to new module
