+++ src/geometry/facets.rs
@@ -1,11 +1,14 @@
 //! Facet modelling & related computation
 
-use num::{zero, FromPrimitive};
-
 use crate::{
-    constants::{mesh::N_FACETS_OUT, CustomFloat},
+    constants::{mesh::N_FACETS_OUT, sim::HUGE_FLOAT, CustomFloat},
     data::mc_vector::MCVector,
 };
+use num::{zero, FromPrimitive};
+
+//
+// Distance to facet
+//
 
 /// Structure used to represent the distance to a given facet.
 ///
@@ -21,6 +24,10 @@ pub struct MCDistanceToFacet<T: CustomFloat> {
     /// Index of the given sub-facet.
     pub subfacet: usize,
 }
+
+//
+// Spatial plane
+//
 
 /// Structure representing a plane of equation `a*x + b*y + c*z + d = 0`
 ///
@@ -67,3 +74,29 @@ impl<T: CustomFloat> MCGeneralPlane<T> {
 
 /// List of planes associated with the outward-facing facets of a cell.
 pub type MCFacetGeometryCell<T> = [MCGeneralPlane<T>; N_FACETS_OUT];
+
+//
+// Nearest facet
+//
+
+/// Structure used to represent the nearest facet to a particle,
+/// holding relevant data for computation.
+#[derive(Debug, Clone, Copy)]
+pub struct MCNearestFacet<T: CustomFloat> {
+    /// Index of the facet the particle is the closest to.
+    pub facet: usize,
+    /// Distance between facet and particle.
+    pub distance_to_facet: T,
+    /// Dot product between facet and direction vector.
+    pub dot_product: T,
+}
+
+impl<T: CustomFloat> Default for MCNearestFacet<T> {
+    fn default() -> Self {
+        Self {
+            facet: 0,
+            distance_to_facet: FromPrimitive::from_f64(HUGE_FLOAT).unwrap(),
+            dot_product: zero(),
+        }
+    }
+}
